requirements.txt
# AutoGen Framework Dependencies
autogen-agentchat>=0.5.7
autogen-core>=0.5.7
autogen-ext[openai]>=0.5.7

# Google Gemini API Support
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.60.0

# Core Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Async Support
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Logging and Monitoring
structlog>=23.2.0
rich>=13.7.0
